+++ test/vec_mat_test.cpp
@@ -74,23 +74,5 @@ TEST(VecMatTest, VecMatOperatorTest2) {
 
 TEST(VecMatTest, VecMatOperatorTest3) {
     PANSFE::Vec<int> a = {1, 2, 3}, b = {4, 5, 6};
-    ASSERT_EQ(a.Transpose() * b, 32);
+    ASSERT_EQ(int(a.Transpose() * b), 32);
 }
-
-// using namespace PANSFE;
-
-// int main() {
-//     Mat<int> A = {{1, 2, 3}, {4, 5, 6}, {7, 8, 9}};
-//     Vec<int> b = {1, 2, 3};
-//     std::cout << A * b << std::endl;
-
-//     Mat<int> B = {{1}, {3}, {5}};
-//     Vec<int> c = A * B;
-//     std::cout << c << std::endl;
-
-//     int d = B.Transpose() * B;
-//     std::cout << d << std::endl;
-
-//     std::cout << A.Hstack(b) << std::endl;
-//     std::cout << b.Hstack(b) << std::endl;
-// }
